# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Talent Management System",
        "description": "TMS Employee Resource Skills Tracking Management System"
    },
    "paths": {
...
        "/deleteSubSkillCategoryByAdmin": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "deleteSubSkillCategoryByAdmin",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "deleteSubSkillCategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deleteSubSkillCategory"
                        }
                    }

                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/deleteStatusByAdmin": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "deleteStatusByAdmin",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "deleteStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deleteStatus"
                        }
                    }

                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "deleteSubSkillCategory": {
            "type": "object",
            "properties": {
                "SkillId": {
                    "format": "int32",
                    "type": "integer"
                },
                "ModifiedBy": {
                    "type": "string"
                },
                "ModifiedDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "deleteStatus": {
            "type": "object",
            "properties": {
                "Id": {
                    "format": "int32",
                    "type": "integer"
                },
                "ModifiedBy": {
                    "type": "string"
                },
                "ModifiedDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
...
    }
}